Reduce duplicate custom scalars tests between core APIs and React hooks
