+++ src/react/hooks/__tests__/useLazyQuery/customScalars.test.tsx
@@ -5,13 +5,7 @@ import {
 import { delay, of } from "rxjs";
 
 import type { OperationVariables } from "@apollo/client";
-import {
-  ApolloClient,
-  ApolloLink,
-  CombinedGraphQLErrors,
-  gql,
-  NetworkStatus,
-} from "@apollo/client";
+import { ApolloClient, ApolloLink, gql, NetworkStatus } from "@apollo/client";
 import { InMemoryCache } from "@apollo/client/cache";
 import { useLazyQuery } from "@apollo/client/react";
 import {
@@ -19,7 +13,7 @@ import {
   dateScalar,
 } from "@apollo/client/testing/internal";
 
-test("serializes scalar variables used in field arguments", async () => {
+test("serializes parsed scalar values in variables", async () => {
   let requestVariables!: OperationVariables;
 
   const link = new ApolloLink((operation) => {
@@ -125,32 +119,38 @@ test("serializes scalar variables used in field arguments", async () => {
   expect(requestVariables).toStrictEqualTyped({ date: "2026-01-01" });
 });
 
-test("serializes scalar variables used in directive arguments", async () => {
-  let requestVariables!: OperationVariables;
-
-  const link = new ApolloLink((operation) => {
-    requestVariables = operation.variables;
-
-    return of({
-      data: { event: { __typename: "Event", name: "GraphQL Summit" } },
-    }).pipe(delay(20));
-  });
-
-  const client = new ApolloClient({
-    cache: new InMemoryCache({
-      scalars: {
-        Date: dateScalar,
-      },
-    }),
-    link,
-  });
+test("returns parsed scalar fields", async () => {
   const query = gql`
-    query Event($date: Date!) {
-      event @on(date: $date) {
-        name
+    query Event {
+      event {
+        id
+        startDate
       }
     }
   `;
+  const client = new ApolloClient({
+    cache: new InMemoryCache({
+      scalars: { Date: dateScalar },
+      typePolicies: {
+        Event: {
+          fields: {
+            startDate: { scalar: "Date" },
+          },
+        },
+      },
+    }),
+    link: new ApolloLink(() =>
+      of({
+        data: {
+          event: {
+            __typename: "Event",
+            id: "1",
+            startDate: "2026-01-01",
+          },
+        },
+      }).pipe(delay(20))
+    ),
+  });
 
   using _disabledAct = disableActEnvironment();
   const { takeSnapshot, getCurrentSnapshot } = await renderHookToSnapshotStream(
@@ -174,417 +174,6 @@ test("serializes scalar variables used in directive arguments", async () => {
 
   const [execute] = getCurrentSnapshot();
 
-  await expect(
-    execute({
-      variables: {
-        date: new Date(2026, 0, 1),
-      },
-    })
-  ).resolves.toStrictEqualTyped({
-    data: {
-      event: {
-        __typename: "Event",
-        name: "GraphQL Summit",
-      },
-    },
-  });
-
-  {
-    const [, result] = await takeSnapshot();
-
-    expect(result).toStrictEqualTyped({
-      data: undefined,
-      dataState: "empty",
-      called: true,
-      loading: true,
-      networkStatus: NetworkStatus.setVariables,
-      previousData: undefined,
-      variables: {
-        date: new Date(2026, 0, 1),
-      },
-    });
-  }
-
-  {
-    const [, result] = await takeSnapshot();
-
-    expect(result).toStrictEqualTyped({
-      data: {
-        event: {
-          __typename: "Event",
-          name: "GraphQL Summit",
-        },
-      },
-      dataState: "complete",
-      called: true,
-      loading: false,
-      networkStatus: NetworkStatus.ready,
-      previousData: undefined,
-      variables: {
-        date: new Date(2026, 0, 1),
-      },
-    });
-  }
-
-  await expect(takeSnapshot).not.toRerender();
-
-  expect(requestVariables).toStrictEqualTyped({ date: "2026-01-01" });
-});
-
-test("serializes scalar fields in input object variables", async () => {
-  let requestVariables!: OperationVariables;
-
-  const link = new ApolloLink((operation) => {
-    requestVariables = operation.variables;
-
-    return of({
-      data: { event: { __typename: "Event", name: "GraphQL Summit" } },
-    }).pipe(delay(20));
-  });
-
-  const client = new ApolloClient({
-    cache: new InMemoryCache({
-      scalars: {
-        Date: dateScalar,
-      },
-      inputObjects: {
-        EventFilter: {
-          fields: {
-            date: "Date",
-          },
-        },
-      },
-    }),
-    link,
-  });
-  const query = gql`
-    query Event($filter: EventFilter!) {
-      event(filter: $filter) {
-        name
-      }
-    }
-  `;
-
-  using _disabledAct = disableActEnvironment();
-  const { takeSnapshot, getCurrentSnapshot } = await renderHookToSnapshotStream(
-    () => useLazyQuery(query),
-    { wrapper: createClientWrapper(client) }
-  );
-
-  {
-    const [, result] = await takeSnapshot();
-
-    expect(result).toStrictEqualTyped({
-      data: undefined,
-      dataState: "empty",
-      called: false,
-      loading: false,
-      networkStatus: NetworkStatus.ready,
-      previousData: undefined,
-      variables: {},
-    });
-  }
-
-  const [execute] = getCurrentSnapshot();
-
-  await expect(
-    execute({
-      variables: {
-        filter: {
-          date: new Date(2026, 0, 1),
-        },
-      },
-    })
-  ).resolves.toStrictEqualTyped({
-    data: {
-      event: {
-        __typename: "Event",
-        name: "GraphQL Summit",
-      },
-    },
-  });
-
-  {
-    const [, result] = await takeSnapshot();
-
-    expect(result).toStrictEqualTyped({
-      data: undefined,
-      dataState: "empty",
-      called: true,
-      loading: true,
-      networkStatus: NetworkStatus.setVariables,
-      previousData: undefined,
-      variables: {
-        filter: {
-          date: new Date(2026, 0, 1),
-        },
-      },
-    });
-  }
-
-  {
-    const [, result] = await takeSnapshot();
-
-    expect(result).toStrictEqualTyped({
-      data: {
-        event: {
-          __typename: "Event",
-          name: "GraphQL Summit",
-        },
-      },
-      dataState: "complete",
-      called: true,
-      loading: false,
-      networkStatus: NetworkStatus.ready,
-      previousData: undefined,
-      variables: {
-        filter: {
-          date: new Date(2026, 0, 1),
-        },
-      },
-    });
-  }
-
-  await expect(takeSnapshot).not.toRerender();
-
-  expect(requestVariables).toStrictEqualTyped({
-    filter: {
-      date: "2026-01-01",
-    },
-  });
-});
-
-test("parses cached custom scalar fields with a cache-only fetch policy", async () => {
-  const query = gql`
-    query Event {
-      event {
-        id
-        startDate
-      }
-    }
-  `;
-  const client = new ApolloClient({
-    cache: new InMemoryCache({
-      scalars: { Date: dateScalar },
-      typePolicies: {
-        Event: {
-          fields: {
-            startDate: { scalar: "Date" },
-          },
-        },
-      },
-    }),
-    link: ApolloLink.empty(),
-  });
-  client.writeQuery({
-    query,
-    data: {
-      event: {
-        __typename: "Event",
-        id: "1",
-        startDate: "2026-01-01",
-      },
-    },
-  });
-
-  using _disabledAct = disableActEnvironment();
-  const { takeSnapshot, getCurrentSnapshot } = await renderHookToSnapshotStream(
-    () => useLazyQuery(query, { fetchPolicy: "cache-only" }),
-    { wrapper: createClientWrapper(client) }
-  );
-
-  {
-    const [, result] = await takeSnapshot();
-
-    expect(result).toStrictEqualTyped({
-      data: undefined,
-      dataState: "empty",
-      called: false,
-      loading: false,
-      networkStatus: NetworkStatus.ready,
-      previousData: undefined,
-      variables: {},
-    });
-  }
-
-  const [execute] = getCurrentSnapshot();
-
-  await expect(execute()).resolves.toStrictEqualTyped({
-    data: {
-      event: {
-        __typename: "Event",
-        id: "1",
-        startDate: new Date(2026, 0, 1),
-      },
-    },
-  });
-
-  {
-    const [, result] = await takeSnapshot();
-
-    expect(result).toStrictEqualTyped({
-      data: {
-        event: {
-          __typename: "Event",
-          id: "1",
-          startDate: new Date(2026, 0, 1),
-        },
-      },
-      dataState: "complete",
-      called: true,
-      loading: false,
-      networkStatus: NetworkStatus.ready,
-      previousData: undefined,
-      variables: {},
-    });
-  }
-
-  await expect(takeSnapshot).not.toRerender();
-});
-
-test("parses cached custom scalar fields with a cache-first fetch policy", async () => {
-  const query = gql`
-    query Event {
-      event {
-        id
-        startDate
-      }
-    }
-  `;
-  const client = new ApolloClient({
-    cache: new InMemoryCache({
-      scalars: { Date: dateScalar },
-      typePolicies: {
-        Event: {
-          fields: {
-            startDate: { scalar: "Date" },
-          },
-        },
-      },
-    }),
-    link: ApolloLink.empty(),
-  });
-  client.writeQuery({
-    query,
-    data: {
-      event: {
-        __typename: "Event",
-        id: "1",
-        startDate: "2026-01-01",
-      },
-    },
-  });
-
-  using _disabledAct = disableActEnvironment();
-  const { takeSnapshot, getCurrentSnapshot } = await renderHookToSnapshotStream(
-    () => useLazyQuery(query, { fetchPolicy: "cache-first" }),
-    { wrapper: createClientWrapper(client) }
-  );
-
-  {
-    const [, result] = await takeSnapshot();
-
-    expect(result).toStrictEqualTyped({
-      data: undefined,
-      dataState: "empty",
-      called: false,
-      loading: false,
-      networkStatus: NetworkStatus.ready,
-      previousData: undefined,
-      variables: {},
-    });
-  }
-
-  const [execute] = getCurrentSnapshot();
-
-  await expect(execute()).resolves.toStrictEqualTyped({
-    data: {
-      event: {
-        __typename: "Event",
-        id: "1",
-        startDate: new Date(2026, 0, 1),
-      },
-    },
-  });
-
-  {
-    const [, result] = await takeSnapshot();
-
-    expect(result).toStrictEqualTyped({
-      data: {
-        event: {
-          __typename: "Event",
-          id: "1",
-          startDate: new Date(2026, 0, 1),
-        },
-      },
-      dataState: "complete",
-      called: true,
-      loading: false,
-      networkStatus: NetworkStatus.ready,
-      previousData: undefined,
-      variables: {},
-    });
-  }
-
-  await expect(takeSnapshot).not.toRerender();
-});
-
-test("parses network custom scalar fields with a cache-first fetch policy", async () => {
-  const query = gql`
-    query Event {
-      event {
-        id
-        startDate
-      }
-    }
-  `;
-  const client = new ApolloClient({
-    cache: new InMemoryCache({
-      scalars: { Date: dateScalar },
-      typePolicies: {
-        Event: {
-          fields: {
-            startDate: { scalar: "Date" },
-          },
-        },
-      },
-    }),
-    link: new ApolloLink(() =>
-      of({
-        data: {
-          event: {
-            __typename: "Event",
-            id: "1",
-            startDate: "2026-01-01",
-          },
-        },
-      }).pipe(delay(20))
-    ),
-  });
-
-  using _disabledAct = disableActEnvironment();
-  const { takeSnapshot, getCurrentSnapshot } = await renderHookToSnapshotStream(
-    () => useLazyQuery(query, { fetchPolicy: "cache-first" }),
-    { wrapper: createClientWrapper(client) }
-  );
-
-  {
-    const [, result] = await takeSnapshot();
-
-    expect(result).toStrictEqualTyped({
-      data: undefined,
-      dataState: "empty",
-      called: false,
-      loading: false,
-      networkStatus: NetworkStatus.ready,
-      previousData: undefined,
-      variables: {},
-    });
-  }
-
-  const [execute] = getCurrentSnapshot();
-
   await expect(execute()).resolves.toStrictEqualTyped({
     data: {
       event: {
@@ -632,438 +221,7 @@ test("parses network custom scalar fields with a cache-first fetch policy", asyn
   await expect(takeSnapshot).not.toRerender();
 });
 
-test("parses cached and network custom scalar fields with a cache-and-network fetch policy", async () => {
-  const query = gql`
-    query Event {
-      event {
-        id
-        startDate
-      }
-    }
-  `;
-  const client = new ApolloClient({
-    cache: new InMemoryCache({
-      scalars: { Date: dateScalar },
-      typePolicies: {
-        Event: {
-          fields: {
-            startDate: { scalar: "Date" },
-          },
-        },
-      },
-    }),
-    link: new ApolloLink(() =>
-      of({
-        data: {
-          event: {
-            __typename: "Event",
-            id: "1",
-            startDate: "2026-02-02",
-          },
-        },
-      }).pipe(delay(20))
-    ),
-  });
-  client.writeQuery({
-    query,
-    data: {
-      event: {
-        __typename: "Event",
-        id: "1",
-        startDate: "2026-01-01",
-      },
-    },
-  });
-
-  using _disabledAct = disableActEnvironment();
-  const { takeSnapshot, getCurrentSnapshot } = await renderHookToSnapshotStream(
-    () => useLazyQuery(query, { fetchPolicy: "cache-and-network" }),
-    { wrapper: createClientWrapper(client) }
-  );
-
-  {
-    const [, result] = await takeSnapshot();
-
-    expect(result).toStrictEqualTyped({
-      data: undefined,
-      dataState: "empty",
-      called: false,
-      loading: false,
-      networkStatus: NetworkStatus.ready,
-      previousData: undefined,
-      variables: {},
-    });
-  }
-
-  const [execute] = getCurrentSnapshot();
-
-  await expect(execute()).resolves.toStrictEqualTyped({
-    data: {
-      event: {
-        __typename: "Event",
-        id: "1",
-        startDate: new Date(2026, 1, 2),
-      },
-    },
-  });
-
-  {
-    const [, result] = await takeSnapshot();
-
-    expect(result).toStrictEqualTyped({
-      data: {
-        event: {
-          __typename: "Event",
-          id: "1",
-          startDate: new Date(2026, 0, 1),
-        },
-      },
-      dataState: "complete",
-      called: true,
-      loading: true,
-      networkStatus: NetworkStatus.loading,
-      previousData: undefined,
-      variables: {},
-    });
-  }
-
-  {
-    const [, result] = await takeSnapshot();
-
-    expect(result).toStrictEqualTyped({
-      data: {
-        event: {
-          __typename: "Event",
-          id: "1",
-          startDate: new Date(2026, 1, 2),
-        },
-      },
-      dataState: "complete",
-      called: true,
-      loading: false,
-      networkStatus: NetworkStatus.ready,
-      previousData: {
-        event: {
-          __typename: "Event",
-          id: "1",
-          startDate: new Date(2026, 0, 1),
-        },
-      },
-      variables: {},
-    });
-  }
-
-  await expect(takeSnapshot).not.toRerender();
-});
-
-test("parses network custom scalar fields with a cache-and-network fetch policy", async () => {
-  const query = gql`
-    query Event {
-      event {
-        id
-        startDate
-      }
-    }
-  `;
-  const client = new ApolloClient({
-    cache: new InMemoryCache({
-      scalars: { Date: dateScalar },
-      typePolicies: {
-        Event: {
-          fields: {
-            startDate: { scalar: "Date" },
-          },
-        },
-      },
-    }),
-    link: new ApolloLink(() =>
-      of({
-        data: {
-          event: {
-            __typename: "Event",
-            id: "1",
-            startDate: "2026-01-01",
-          },
-        },
-      }).pipe(delay(20))
-    ),
-  });
-
-  using _disabledAct = disableActEnvironment();
-  const { takeSnapshot, getCurrentSnapshot } = await renderHookToSnapshotStream(
-    () => useLazyQuery(query, { fetchPolicy: "cache-and-network" }),
-    { wrapper: createClientWrapper(client) }
-  );
-
-  {
-    const [, result] = await takeSnapshot();
-
-    expect(result).toStrictEqualTyped({
-      data: undefined,
-      dataState: "empty",
-      called: false,
-      loading: false,
-      networkStatus: NetworkStatus.ready,
-      previousData: undefined,
-      variables: {},
-    });
-  }
-
-  const [execute] = getCurrentSnapshot();
-
-  await expect(execute()).resolves.toStrictEqualTyped({
-    data: {
-      event: {
-        __typename: "Event",
-        id: "1",
-        startDate: new Date(2026, 0, 1),
-      },
-    },
-  });
-
-  {
-    const [, result] = await takeSnapshot();
-
-    expect(result).toStrictEqualTyped({
-      data: undefined,
-      dataState: "empty",
-      called: true,
-      loading: true,
-      networkStatus: NetworkStatus.loading,
-      previousData: undefined,
-      variables: {},
-    });
-  }
-
-  {
-    const [, result] = await takeSnapshot();
-
-    expect(result).toStrictEqualTyped({
-      data: {
-        event: {
-          __typename: "Event",
-          id: "1",
-          startDate: new Date(2026, 0, 1),
-        },
-      },
-      dataState: "complete",
-      called: true,
-      loading: false,
-      networkStatus: NetworkStatus.ready,
-      previousData: undefined,
-      variables: {},
-    });
-  }
-
-  await expect(takeSnapshot).not.toRerender();
-});
-
-test("parses network custom scalar fields with a network-only fetch policy", async () => {
-  const query = gql`
-    query Event {
-      event {
-        id
-        startDate
-      }
-    }
-  `;
-  const client = new ApolloClient({
-    cache: new InMemoryCache({
-      scalars: { Date: dateScalar },
-      typePolicies: {
-        Event: {
-          fields: {
-            startDate: { scalar: "Date" },
-          },
-        },
-      },
-    }),
-    link: new ApolloLink(() =>
-      of({
-        data: {
-          event: {
-            __typename: "Event",
-            id: "1",
-            startDate: "2026-01-01",
-          },
-        },
-      }).pipe(delay(20))
-    ),
-  });
-
-  using _disabledAct = disableActEnvironment();
-  const { takeSnapshot, getCurrentSnapshot } = await renderHookToSnapshotStream(
-    () => useLazyQuery(query, { fetchPolicy: "network-only" }),
-    { wrapper: createClientWrapper(client) }
-  );
-
-  {
-    const [, result] = await takeSnapshot();
-
-    expect(result).toStrictEqualTyped({
-      data: undefined,
-      dataState: "empty",
-      called: false,
-      loading: false,
-      networkStatus: NetworkStatus.ready,
-      previousData: undefined,
-      variables: {},
-    });
-  }
-
-  const [execute] = getCurrentSnapshot();
-
-  await expect(execute()).resolves.toStrictEqualTyped({
-    data: {
-      event: {
-        __typename: "Event",
-        id: "1",
-        startDate: new Date(2026, 0, 1),
-      },
-    },
-  });
-
-  {
-    const [, result] = await takeSnapshot();
-
-    expect(result).toStrictEqualTyped({
-      data: undefined,
-      dataState: "empty",
-      called: true,
-      loading: true,
-      networkStatus: NetworkStatus.loading,
-      previousData: undefined,
-      variables: {},
-    });
-  }
-
-  {
-    const [, result] = await takeSnapshot();
-
-    expect(result).toStrictEqualTyped({
-      data: {
-        event: {
-          __typename: "Event",
-          id: "1",
-          startDate: new Date(2026, 0, 1),
-        },
-      },
-      dataState: "complete",
-      called: true,
-      loading: false,
-      networkStatus: NetworkStatus.ready,
-      previousData: undefined,
-      variables: {},
-    });
-  }
-
-  await expect(takeSnapshot).not.toRerender();
-});
-
-test.failing(
-  "parses custom scalar fields with a no-cache fetch policy",
-  async () => {
-    const query = gql`
-      query Event {
-        event {
-          id
-          startDate
-        }
-      }
-    `;
-    const client = new ApolloClient({
-      cache: new InMemoryCache({
-        scalars: { Date: dateScalar },
-        typePolicies: {
-          Event: {
-            fields: {
-              startDate: { scalar: "Date" },
-            },
-          },
-        },
-      }),
-      link: new ApolloLink(() =>
-        of({
-          data: {
-            event: {
-              __typename: "Event",
-              id: "1",
-              startDate: "2026-01-01",
-            },
-          },
-        }).pipe(delay(20))
-      ),
-    });
-
-    using _disabledAct = disableActEnvironment();
-    const { takeSnapshot, getCurrentSnapshot } =
-      await renderHookToSnapshotStream(
-        () => useLazyQuery(query, { fetchPolicy: "no-cache" }),
-        { wrapper: createClientWrapper(client) }
-      );
-
-    {
-      const [, result] = await takeSnapshot();
-
-      expect(result).toStrictEqualTyped({
-        data: undefined,
-        dataState: "empty",
-        called: false,
-        loading: false,
-        networkStatus: NetworkStatus.ready,
-        previousData: undefined,
-        variables: {},
-      });
-    }
-    const [execute] = getCurrentSnapshot();
-
-    await expect(execute()).resolves.toStrictEqualTyped({
-      data: {
-        event: {
-          __typename: "Event",
-          id: "1",
-          startDate: new Date(2026, 0, 1),
-        },
-      },
-    });
-
-    {
-      const [, result] = await takeSnapshot();
-
-      expect(result).toStrictEqualTyped({
-        data: undefined,
-        dataState: "empty",
-        called: true,
-        loading: true,
-        networkStatus: NetworkStatus.loading,
-        previousData: undefined,
-        variables: {},
-      });
-    }
-
-    {
-      const [, result] = await takeSnapshot();
-
-      expect(result).toStrictEqualTyped({
-        data: {
-          event: {
-            __typename: "Event",
-            id: "1",
-            startDate: new Date(2026, 0, 1),
-          },
-        },
-        dataState: "complete",
-        called: true,
-        loading: false,
-        networkStatus: NetworkStatus.ready,
-        previousData: undefined,
-        variables: {},
-      });
-    }
-  }
-);
-
-test("preserves referential identity when re-executing with identical serialized scalar values", async () => {
+test("preserves referential identity when re-executing with identical scalar values", async () => {
   const query = gql`
     query Event {
       event {
@@ -1215,396 +373,6 @@ test("preserves referential identity when re-executing with identical serialized
     });
 
     expect(result.data).toBe(previousData);
-  }
-
-  await expect(takeSnapshot).not.toRerender();
-});
-
-test("serializes scalar fields in the error with a `none` error policy", async () => {
-  const query = gql`
-    query Event {
-      event {
-        id
-        startDate
-        endDate
-      }
-    }
-  `;
-  const client = new ApolloClient({
-    cache: new InMemoryCache({
-      scalars: { Date: dateScalar },
-      typePolicies: {
-        Event: {
-          fields: {
-            startDate: { scalar: "Date" },
-            endDate: { scalar: "Date" },
-          },
-        },
-      },
-    }),
-    link: new ApolloLink(() =>
-      of({
-        data: {
-          event: {
-            __typename: "Event",
-            id: "1",
-            startDate: "2026-01-01",
-            endDate: null,
-          },
-        },
-        errors: [
-          {
-            message: "Could not resolve endDate",
-            path: ["event", "endDate"],
-          },
-        ],
-      }).pipe(delay(20))
-    ),
-  });
-
-  using _disabledAct = disableActEnvironment();
-  const { takeSnapshot, getCurrentSnapshot } = await renderHookToSnapshotStream(
-    () => useLazyQuery(query, { errorPolicy: "none" }),
-    { wrapper: createClientWrapper(client) }
-  );
-
-  {
-    const [, result] = await takeSnapshot();
-
-    expect(result).toStrictEqualTyped({
-      data: undefined,
-      dataState: "empty",
-      called: false,
-      loading: false,
-      networkStatus: NetworkStatus.ready,
-      previousData: undefined,
-      variables: {},
-    });
-  }
-
-  const [execute] = getCurrentSnapshot();
-
-  await expect(execute()).rejects.toEqual(
-    new CombinedGraphQLErrors({
-      data: {
-        event: {
-          __typename: "Event",
-          id: "1",
-          startDate: "2026-01-01",
-          endDate: null,
-        },
-      },
-      errors: [
-        {
-          message: "Could not resolve endDate",
-          path: ["event", "endDate"],
-        },
-      ],
-    })
-  );
-
-  {
-    const [, result] = await takeSnapshot();
-
-    expect(result).toStrictEqualTyped({
-      data: undefined,
-      dataState: "empty",
-      called: true,
-      loading: true,
-      networkStatus: NetworkStatus.loading,
-      previousData: undefined,
-      variables: {},
-    });
-  }
-
-  {
-    const [, result] = await takeSnapshot();
-
-    expect(result).toStrictEqualTyped({
-      data: undefined,
-      dataState: "empty",
-      error: new CombinedGraphQLErrors({
-        data: {
-          event: {
-            __typename: "Event",
-            id: "1",
-            startDate: "2026-01-01",
-            endDate: null,
-          },
-        },
-        errors: [
-          {
-            message: "Could not resolve endDate",
-            path: ["event", "endDate"],
-          },
-        ],
-      }),
-      called: true,
-      loading: false,
-      networkStatus: NetworkStatus.error,
-      previousData: undefined,
-      variables: {},
-    });
-  }
-
-  await expect(takeSnapshot).not.toRerender();
-});
-
-test("parses scalar fields in the result and serializes them in the error with an `all` error policy", async () => {
-  const query = gql`
-    query Event {
-      event {
-        id
-        startDate
-        endDate
-      }
-    }
-  `;
-  const client = new ApolloClient({
-    cache: new InMemoryCache({
-      scalars: { Date: dateScalar },
-      typePolicies: {
-        Event: {
-          fields: {
-            startDate: { scalar: "Date" },
-            endDate: { scalar: "Date" },
-          },
-        },
-      },
-    }),
-    link: new ApolloLink(() =>
-      of({
-        data: {
-          event: {
-            __typename: "Event",
-            id: "1",
-            startDate: "2026-01-01",
-            endDate: null,
-          },
-        },
-        errors: [
-          {
-            message: "Could not resolve endDate",
-            path: ["event", "endDate"],
-          },
-        ],
-      }).pipe(delay(20))
-    ),
-  });
-
-  using _disabledAct = disableActEnvironment();
-  const { takeSnapshot, getCurrentSnapshot } = await renderHookToSnapshotStream(
-    () => useLazyQuery(query, { errorPolicy: "all" }),
-    { wrapper: createClientWrapper(client) }
-  );
-
-  {
-    const [, result] = await takeSnapshot();
-
-    expect(result).toStrictEqualTyped({
-      data: undefined,
-      dataState: "empty",
-      called: false,
-      loading: false,
-      networkStatus: NetworkStatus.ready,
-      previousData: undefined,
-      variables: {},
-    });
-  }
-
-  const [execute] = getCurrentSnapshot();
-
-  await expect(execute()).resolves.toStrictEqualTyped({
-    data: {
-      event: {
-        __typename: "Event",
-        id: "1",
-        startDate: new Date(2026, 0, 1),
-        endDate: null,
-      },
-    },
-    error: new CombinedGraphQLErrors({
-      data: {
-        event: {
-          __typename: "Event",
-          id: "1",
-          // TODO: Determine if this is correct
-          startDate: new Date(2026, 0, 1),
-          endDate: null,
-        },
-      },
-      errors: [
-        {
-          message: "Could not resolve endDate",
-          path: ["event", "endDate"],
-        },
-      ],
-    }),
-  });
-
-  {
-    const [, result] = await takeSnapshot();
-
-    expect(result).toStrictEqualTyped({
-      data: undefined,
-      dataState: "empty",
-      called: true,
-      loading: true,
-      networkStatus: NetworkStatus.loading,
-      previousData: undefined,
-      variables: {},
-    });
-  }
-
-  {
-    const [, result] = await takeSnapshot();
-
-    expect(result).toStrictEqualTyped({
-      data: {
-        event: {
-          __typename: "Event",
-          id: "1",
-          startDate: new Date(2026, 0, 1),
-          endDate: null,
-        },
-      },
-      dataState: "complete",
-      error: new CombinedGraphQLErrors({
-        data: {
-          event: {
-            __typename: "Event",
-            id: "1",
-            // TODO: Determine if this is correct
-            startDate: new Date(2026, 0, 1),
-            endDate: null,
-          },
-        },
-        errors: [
-          {
-            message: "Could not resolve endDate",
-            path: ["event", "endDate"],
-          },
-        ],
-      }),
-      called: true,
-      loading: false,
-      networkStatus: NetworkStatus.error,
-      previousData: undefined,
-      variables: {},
-    });
-  }
-
-  await expect(takeSnapshot).not.toRerender();
-});
-
-test("parses custom scalar fields with an `ignore` error policy", async () => {
-  const query = gql`
-    query Event {
-      event {
-        id
-        startDate
-        endDate
-      }
-    }
-  `;
-  const client = new ApolloClient({
-    cache: new InMemoryCache({
-      scalars: { Date: dateScalar },
-      typePolicies: {
-        Event: {
-          fields: {
-            startDate: { scalar: "Date" },
-            endDate: { scalar: "Date" },
-          },
-        },
-      },
-    }),
-    link: new ApolloLink(() =>
-      of({
-        data: {
-          event: {
-            __typename: "Event",
-            id: "1",
-            startDate: "2026-01-01",
-            endDate: null,
-          },
-        },
-        errors: [
-          {
-            message: "Could not resolve endDate",
-            path: ["event", "endDate"],
-          },
-        ],
-      }).pipe(delay(20))
-    ),
-  });
-
-  using _disabledAct = disableActEnvironment();
-  const { takeSnapshot, getCurrentSnapshot } = await renderHookToSnapshotStream(
-    () => useLazyQuery(query, { errorPolicy: "ignore" }),
-    { wrapper: createClientWrapper(client) }
-  );
-
-  {
-    const [, result] = await takeSnapshot();
-
-    expect(result).toStrictEqualTyped({
-      data: undefined,
-      dataState: "empty",
-      called: false,
-      loading: false,
-      networkStatus: NetworkStatus.ready,
-      previousData: undefined,
-      variables: {},
-    });
-  }
-
-  const [execute] = getCurrentSnapshot();
-
-  await expect(execute()).resolves.toStrictEqualTyped({
-    data: {
-      event: {
-        __typename: "Event",
-        id: "1",
-        startDate: new Date(2026, 0, 1),
-        endDate: null,
-      },
-    },
-  });
-
-  {
-    const [, result] = await takeSnapshot();
-
-    expect(result).toStrictEqualTyped({
-      data: undefined,
-      dataState: "empty",
-      called: true,
-      loading: true,
-      networkStatus: NetworkStatus.loading,
-      previousData: undefined,
-      variables: {},
-    });
-  }
-
-  {
-    const [, result] = await takeSnapshot();
-
-    expect(result).toStrictEqualTyped({
-      data: {
-        event: {
-          __typename: "Event",
-          id: "1",
-          startDate: new Date(2026, 0, 1),
-          endDate: null,
-        },
-      },
-      dataState: "complete",
-      called: true,
-      loading: false,
-      networkStatus: NetworkStatus.ready,
-      previousData: undefined,
-      variables: {},
-    });
   }
 
   await expect(takeSnapshot).not.toRerender();
